Extract base ModuleRecord from CyclicModuleRecord to match ecmascript specs
